requirements.txt
numpy
opencv-python
pyautogui
PySimpleGUI-4-foss
